fintech-lms/lms_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-cors-headers
python-dotenv
dj-database-url
whitenoise
